make a new class for the functions and move them to the new class
