+++ Src/BootCamp.Chapter/Program.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.IO;
-using Console = System.Console;
+﻿using Console = System.Console;
 
 namespace BootCamp.Chapter
 {
@@ -10,33 +6,8 @@ namespace BootCamp.Chapter
     {
         static void Main(string[] args)
         {
-            var test = GetFloat("give me your height in metres: "); 
+            Class3.Demo();  
         }
-
-        public static float GetBmi(float weight, float height)
-        {
-            return weight / (float)Math.Pow(height, 2);
-        }
-
-        public static string GetString(string message)
-        {
-            Console.Write(message);
-            var input = Console.ReadLine();
-            return input; 
-        }
-
-        public static int GetInt(string message)
-        {
-            Console.Write(message);
-            var input = int.Parse(Console.ReadLine());
-            return input; 
-        }
-
-        public static float GetFloat(string message)
-        {
-            Console.Write(message);
-            var input = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture); 
-            return input;
-        }
+      
     }
 }
